+++ Apptentive/Apptentive/ApptentiveStyleSheet.m
@@ -7,7 +7,7 @@
 //
 
 #import "ApptentiveStyleSheet.h"
-#import "Apptentive.h"
+#import "ApptentiveMain.h"
 
 NS_ASSUME_NONNULL_BEGIN
