+++ Apptentive/ApptentiveTests/ApptentiveTests-Bridging-Header.h
@@ -2,7 +2,7 @@
 //  Use this file to import your target's public headers that you would like to expose to Swift.
 //
 
-#import "Apptentive.h"
+#import "ApptentiveMain.h"
 #import "ApptentiveLog.h"
 
 #import "ApptentiveConversationManager.h"
