+++ Apptentive/ApptentiveTests/Info.plist
@@ -15,7 +15,7 @@
 	<key>CFBundlePackageType</key>
 	<string>BNDL</string>
 	<key>CFBundleShortVersionString</key>
-	<string>5.2.14</string>
+	<string>5.3.0</string>
 	<key>CFBundleVersion</key>
 	<string>1</string>
 </dict>
