+++ Example/podfile
@@ -1,4 +1,4 @@
-platform :ios, '9.0'
+platform :ios, '10.3'
 use_frameworks!
 
 target 'iOSExample' do
